version 1.4

+++ src/main/java/com/hrsystem/performance/controller/PerformanceTempletController.java
@@ -1,5 +1,7 @@
 package com.hrsystem.performance.controller;
 
+import java.util.List;
+
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,8 +14,11 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+
 import com.hrsystem.common.ExtjsPageRequest;
+import com.hrsystem.performance.entity.Performance;
 import com.hrsystem.performance.entity.PerformanceTemplet;
+import com.hrsystem.performance.service.IPerformanceService;
 import com.hrsystem.performance.service.IPerformanceTempletService;
 
 @RestController
@@ -22,6 +27,8 @@ public class PerformanceTempletController {
 	@Autowired
 	private IPerformanceTempletService performanceTempletService;
 	
+	@Autowired
+	private IPerformanceService performanceService;
 	/**
 	 * 1、查
 	 * @param id
@@ -41,9 +48,9 @@ public class PerformanceTempletController {
 	public String insertPerformance(PerformanceTemplet performanceTemplet) {		
 		try {
 			performanceTempletService.insertPerformanceTemplet(performanceTemplet);
-			return "success:true";
+			return "success:添加成功";
 		} catch (Exception e) {
-			return "success:false";
+			return "success:添加失败";
 		}
 	}
 
@@ -54,12 +61,13 @@ public class PerformanceTempletController {
 	 */
 	@DeleteMapping("/delete/{id}")
 	public String deletePerformanceById(@PathVariable Long id) {
-		try {
-			performanceTempletService.deletePerformanceTemplet(id);
-			return "success:true";
-		} catch (Exception e) {
-			return "success:false";
-		}
+			List<Performance> performanceByPerformanceTempletId = performanceService.getPerformanceByPerformanceTempletId(id);
+			if (performanceByPerformanceTempletId.size() == 0) {
+				performanceTempletService.deletePerformanceTemplet(id);
+				return "success:删除成功";
+			}else {
+				return "success:不能被删除，模板正在被使用";
+			}
 	}
 
 	@GetMapping
